refactor(core): give the walk bound and the typed-write verdict one home
Two rules the backends were each holding a copy of.

`may_explore_children` was byte-identical in the AT-SPI, UIA and AX readers.
Its own doc argues the case: a reader's `walk` and `find_nth` assign ids by
arrival order and `set_value` re-walks to a caller-supplied id, so a bound
applied in one traversal and not the other resolves that id against a
different tree. It reads and records nothing but the budget, so it is now a
`WalkBudget` method.

`verify_write` — 94 lines judging a typed write from the tree read back after
it — was duplicated between the Android `uiautomator` reader and the iOS
reader, differing only in two comment lines, with a "Twin of … keep the two in
step" note on each standing in for the sharing. It composes `AxTarget`,
`Located` and the typed-write predicates, all of which are glass-core's, so it
moves there as `verify_typed_write` alongside them, taking
`TAP_MAY_HAVE_MISSED` and `read_back_failed` with it.

The tests move with the rules. Both backends had grown their own suite for the
verdict: 18 cases under the same name, and one platform's tree fixtures behind
each. Core now holds the union — the two the readers covered separately (an
unreadable subtree hiding a second match, a mask reformatting what it was
given) included — so the branches neither side had reached are covered once,
and the mutation gate has killers in the crate that owns the code.

No behaviour change: the verdicts, their messages and the order they are
decided in are what they were.
