+++ crates/glass-core/src/accessibility.rs
@@ -509,6 +509,31 @@ impl WalkBudget {
         depth >= self.limits.depth
     }
 
+    /// Whether this node's children may be explored, recording the bound that stopped the walk
+    /// when they may not. Callers only consult this once they already know the child list is
+    /// non-empty — calling it for a childless node would record a truncation for declining to
+    /// explore a list that was never going to be walked anyway.
+    ///
+    /// Every traversal that assigns ids MUST reach its decision through this one method, at the
+    /// same point in the walk. A reader's `walk` and `find_nth` number nodes by arrival order and
+    /// `set_value` re-walks to a caller-supplied id, so a bound applied in one traversal but not
+    /// the other resolves that id against a different tree and writes to the wrong element.
+    //
+    // The exactly-at-cap boundary — a *complete* tree of exactly `limits.nodes` must report `None`,
+    // since the last node to arrive wasn't declined — is unit-tested in the Android/iOS mappers,
+    // which can size a synthetic tree to the node. A live GTK tree can't be.
+    pub fn may_explore_children(&mut self, depth: usize) -> bool {
+        if self.depth_exhausted(depth) {
+            self.hit(TruncationLimit::Depth);
+            return false;
+        }
+        if self.nodes_exhausted() {
+            self.hit(TruncationLimit::Nodes);
+            return false;
+        }
+        true
+    }
+
     /// Record a subtree dropped because the platform refused to read a child.
     ///
     /// Distinct from [`WalkBudget::hit`]: a bound is a limit glass chose and can raise, while
@@ -2886,6 +2911,49 @@ mod tests {
         assert!(
             b.nodes_exhausted(),
             "the cap is reached at exactly MAX_NODES"
+        );
+    }
+
+    #[test]
+    fn below_the_caps_children_may_be_explored_and_nothing_is_recorded() {
+        let mut b = WalkBudget::new();
+        assert!(b.may_explore_children(0));
+        assert!(b.truncation().is_none());
+    }
+
+    #[test]
+    fn at_max_depth_the_depth_bound_is_recorded_and_children_may_not_be_explored() {
+        let mut b = WalkBudget::new();
+        assert!(!b.may_explore_children(MAX_DEPTH));
+        assert_eq!(
+            b.truncation().map(|t| t.limit),
+            Some(TruncationLimit::Depth)
+        );
+    }
+
+    #[test]
+    fn with_the_node_budget_spent_the_nodes_bound_is_recorded_and_children_may_not_be_explored() {
+        let mut b = WalkBudget::new();
+        for _ in 0..MAX_NODES {
+            b.visit();
+        }
+        assert!(!b.may_explore_children(0));
+        assert_eq!(
+            b.truncation().map(|t| t.limit),
+            Some(TruncationLimit::Nodes)
+        );
+    }
+
+    #[test]
+    fn when_both_bounds_are_exhausted_the_recorded_limit_is_depth() {
+        let mut b = WalkBudget::new();
+        for _ in 0..MAX_NODES {
+            b.visit();
+        }
+        assert!(!b.may_explore_children(MAX_DEPTH));
+        assert_eq!(
+            b.truncation().map(|t| t.limit),
+            Some(TruncationLimit::Depth)
         );
     }
