+++ crates/glass-core/src/lib.rs
@@ -17,7 +17,8 @@ pub use bounded::{run_bounded, run_bounded_until, run_bounded_with_stdin};
 
 pub mod set_value;
 pub use set_value::{
-    read_back_confirms, typed_clear_landed, typed_text_landed, write_took_no_effect,
+    TAP_MAY_HAVE_MISSED, read_back_confirms, read_back_failed, typed_clear_landed,
+    typed_text_landed, verify_typed_write, write_took_no_effect,
 };
 
 pub mod frame;
